rollback entities list changes
